+++ test/util.js
@@ -4,7 +4,7 @@ var fs = require('fs');
 
 module.exports = {
   read: function (fd) {
-    var buf = new Buffer(1024);
+    var buf = Buffer.alloc(1024);
     fs.readSync(fd, buf, 0, buf.length, null);
   }
 };