+++ History.md
@@ -1,3 +1,7 @@
+Unpublished
+===========
+  * replace new Buffer with Buffer.from or Buffer.alloc
+
 2.0.1 - Apr 02 2018
 ===================
   * update dependencies (nan v2.10.0)